updated faulty routes

+++ backend/express_server/src/routes/hasCategories.js
@@ -2,12 +2,18 @@ import Router from "express-promise-router";
 import validate from "../schema/validate.js";
 import hasCategoriesSchema from "../schema/hasCategory-schema.js";
 import hasCategoriesDB from "../db/hasCategories.js";
+import postsDB from "../db/posts.js";
 import { isAuthorized } from "../util/auth.js";
 import { refreshToken } from "./auth.js";
 
 const hasCategories = Router();
 
 hasCategories.post( "/", isAuthorized, validate({ body: hasCategoriesSchema }), refreshToken, async (req, res) => {
+  let post = { status:"", result:"" };
+  post = await postsDB.getPost(req.body.post_id);
+  if (req.id !== post.result.user_id) {
+    return res.sendStatus(401);
+  }
     const { status, result } = await hasCategoriesDB.postHasCategory(
         req.body.post_id,
         req.body.category_id
@@ -22,6 +28,11 @@ hasCategories.post( "/", isAuthorized, validate({ body: hasCategoriesSchema }), 
 );
 
 hasCategories.delete("/", isAuthorized, refreshToken, async (req, res) => {
+  let post = { status:"", result:"" };
+  post = await postsDB.getPost(req.body.post_id);
+  if (req.id !== post.result.user_id) {
+    return res.sendStatus(401);
+  }
   const { status, result } = await hasCategoriesDB.deleteHasCategory(
     req.body.post_id,
     req.body.category_id
